+++ R/light_global_surrogate.R
@@ -48,6 +48,10 @@ light_global_surrogate.flashlight <- function(x, data = x$data, by = x$by,
                                               v = NULL, use_linkinv = TRUE,
                                               n_max = Inf, seed = NULL,
                                               keep_max_levels = 4L, ...) {
+  if (length(by) >= 2L) {
+    message("The use of more than one 'by' variable is deprecated and will be removed in flashlight 0.10.0.")
+  }
+
   warning_on_names(c("label_name", "tree_name"), ...)
 
   label_name <- getOption("flashlight.label_name")