+++ R/light_importance.R
@@ -81,6 +81,10 @@ light_importance.flashlight <- function(x, data = x$data, by = x$by,
                                         use_linkinv = FALSE, ...) {
   type <- match.arg(type)
 
+  if (length(by) >= 2L) {
+    message("The use of more than one 'by' variable is deprecated and will be removed in flashlight 0.10.0.")
+  }
+
   warning_on_names(
     c("metric_name", "value_name", "label_name", "variable_name", "error_name"), ...
   )