+++ R/light_performance.R
@@ -43,6 +43,10 @@ light_performance.default <- function(x, ...) {
 light_performance.flashlight <- function(x, data = x$data, by = x$by,
                                          metrics = x$metrics,
                                          use_linkinv = FALSE, ...) {
+  if (length(by) >= 2L) {
+    message("The use of more than one 'by' variable is deprecated and will be removed in flashlight 0.10.0.")
+  }
+
   warning_on_names(c("metric_name", "value_name", "label_name"), ...)
 
   metric_name <- getOption("flashlight.metric_name")